Structure ses/rook upstream the same way
This adds build/preinstall/install steps to both implementation.
Having a build() method is necessary in ses (though is does nothing
usefull) so we can call it early in threaded tests that will
opportunitely build rook if possible before deployment.

Move install code to base class as it is no upstream specific.

Move ceph features method (rbd, fs...) to base class too.
